r2575/sofa-dev : MODIFY : "State Events" that are notifying the "StateChanges", ie : addition or the removal of mesh vertices. Now the "MechanicalObject" handles itself the addition or the removal of mesh vertices. The Topology objects are now expected to only read the DOFs and never write them.
ADD : class "StateChangeVisitor" to propagate in a specific way the state events by first notifying the "MechanicalObject" if vertices are going to be added, and by last notifying the "MechanicalObject" if vertices are going to be removed.

ADD : access from class "BaseMeshTopology" to the list of TopologicalChanges (later the same will be done for the list of StateChanges) and to the current number of vertices (getDOFNumber() method).

All "dynamic" components which handle topological changes are now prepared to rely on the "BaseMeshTopology" object instead of downcasting TopologyContainers (this will be done later).

FIX : defensive implementation for Shell computation.

 //partial sync (25/30 changes)//