r3296/sofa-dev : ADD: SparseGridMultipleTopology
Build a SparseGridTopology for several given Triangular meshes.

A stiffness coefficient has to be assigned for each mesh. The last found stiffness coefficient is used for an element shared by several meshes => The mesh ordering is important, and so, more specific stiffness informations must appear in last.

In the example file, the frog has very soft lips and very stiff eyes.

 //partial sync (6/7 changes)//