r12240/sofa-dev : Added a command line option in runSofa to switch on the display of computation time, and made the hierarchy more visible in the output.
Created Compliant/examples/CantileverBeam_4_4_20.scn to compare with the assembly and solution used in examples/Components/linearsolver/CSPARSE-SparseCholeskySolver.scn. The new assembly is slower, in the case of FEM without mappings. The result of the simulation is different also, the Flexible FEM have a non-symmetric behavior compared with the classical FEM.

 //partial sync (11/13 changes)//


Former-commit-id: 8d28826e5d95f71ef76ef6ec5b9384415b098e24