r9721/sofa : DAG: fixing Directed Acyclic scene Graph traversal while the graph structure is modified (adding/deleting nodes) during the traversal.
The traversal flags per visitor are now stored as a map on DAGNode* and the scene graph structure (that can change) is now used for the traversal. Rather than performing the traversal on a precomputed, static sub-graph structure that was not considering the scene graph structure changes.

Note: there is still an undetermined case, when a visitor is run from a sub-graph that contains a multi-node linked with a node outside of the sub-graph.

Former-commit-id: 4b29d3d0cae603a282e7b9d9fd82a135d77425ac