r10817/sofa : push 089ba13..0f2f194
* commit 0f2f1948bd93ed341b1b773dc2f526a697af0b8e
Fix suspicious code detected by clang

This fixes a double negation in TriangularBiquadraticSpringsForceField.
I can't pretend I fully understand what I modified, but from to the rest
of the code, it looks like h should be between 0 and 1. The component is
a lot more stable anyway, so I'm pretty confident.

   1 files changed, 1 insertions(+), 2 deletions(-)




Former-commit-id: ba702436c536616e672b5d45decb5219228c3d0f