r10914/sofa : push 52dabe5..9d47a65
* commit 9d47a65da7df092e6ea921f68ac966cab3cc9123
Add and use an 'Index' typedef in BaseVector and BaseMatrix

BaseVector and BaseMatrix use both int and unsigned int as index types,
and (worse), subclasses usually don't distinguish between unsigned int
and size_t. This leads to inconsistencies everywhere, which in turn
leads to many warnings in some circumstances. This commit is a first
step toward correctness.

Subclasses of BaseVector or BaseMatrix usually defines an Index type,
but don't use it consistently. Instead of this, this commit adds an
Index type in BaseVector and BaseMatrix. For now, it is defined as 'int'
and it is used where ints were used. This means that I didn't replace
unsigned int with Index, and that in the end, types have *not* actually
changed yet.

So this commit tries to make things as consistent as possible without
breaking the interface, so that replacing unsigned int with Index and
changing Index from int to std::size_t (because it's the underlying
idea) will be as little painful as possible.

   15 files changed, 772 insertions(+), 772 deletions(-)




Former-commit-id: db873a8bd45fa20190a1909863ae1752481b101e