r11149/sofa : push 6ebd923..a8898dd
* commit a8898dd41b23ecd6e3302d297c41da7a3eca41ed
Flexible: fix compilation with clang

This commit reorders some includes to make sure that libeigen headers
are included after X11 header X.h.

Reason: eigen uses 'Success' as an enum value, and X11 carelessly
defines a 'Success' preprocessor symbol. Eigen tries to detect this,
but for some reason it was not catched in one case. Fortunately, both
values for 'Success' are 0, so it is not a matter of correctness (but
it could be, so it is comforting to know that clang may choke on those
issues. Well played clang, well played.)

   1 files changed, 7 insertions(+), 1 deletions(-)




Former-commit-id: 29eea864f72ac1adf59c8da6f6d2c0952eaba7d7