r11202/sofa : push bdff841..addb812
* commit addb8121c8db870696043316430b738c6b595726
doxygen: generate a page which lists the components in modules/

As a result of the splitting of the documentation in chunks, finding the
documentation of a class is awkward. This commits improves on this on
two points:

- It merges back the documentation of the framework itself with the main
  page, so that one can find a symbol with the search box again.

- It adds a little program that generates a page listing all the
  components available in modules, so that one can quickly find a
  component without knowing in which module it is.

|  8 files changed, 132 insertions(+), 32 deletions(-)

* commit eb5464d34d8a2ad9f9b31facf05a6bd07db4763c
Move framework/sofa/doc.h to framework/

I want doc.h to be preprocessed by the cmake scripts, so I just move it
out of sofa/ to prevent it from being used as is by doxygen.

   2 files changed, 173 insertions(+), 173 deletions(-)




Former-commit-id: 48aea6de925fab344f7c83cd79963c4871e7c4a4