core: add a mechanism similar to RegisterObject for Plugin
Different approach for registering a plugin's components, which keeps
the code responsible for the registration of a component in its .cpp
file. It is done in such a way that it is almost source compatible with
the way components are registered in the ObjectFactory. (Props to
Jeremie for the idea :-)

A plugin uses the following macro to declare a subclass of Plugin:
SOFA_DECL_PLUGIN(MyPlugin)
and the following macro to define its entry point:
SOFA_PLUGIN_ENTRY_POINT(MyPlugin);

Then, calls to RegisterObject() are replaced with calls to
MyPlugin::registerComponent():
-sofa::core::RegisterObject("...")
+MyPlugin::registerComponent("...")

Former-commit-id: f032bdfb2ce8e3a08ea7a676e9957eac45faf1fa