Add .gitattributes file to normalize line endings
To deal with the different line endings used on Windows (CRLF instead of
LF), git can auto-convert line endings if asked politely.

Normalized files will be stored with LF only, and for Windows users,
they will be converted to CRLF on checkout, and converted to LF on
commit.

The .gitattributes file allows to specify at the repository level which
files should be normalized, using the 'text' attribute.


Former-commit-id: 37859363709f2f8f1d10077b1cad9de25376bd0e