Fix: cleaned dangling parent pointers in children when destroying their parent.
NB: it is possible for an external source to keep a strong pointer to a child node past the destruction of it's parent and to call a method that accesses the hierarchy before releasing the node. Currently, this happens when the contact manager cleans the previous contacts at the start of a frame and an object in contact has been destroyed between frames, but a more generic fix seems more appropriate.

Former-commit-id: 295a5ef27e6347b6393e969e08d13ba3b8d7d4af