defaulttype: add AbstractTypeInfo::Container()
DataTypeInfo was lacking a way to properly discriminate simple values
from container types.  For example, it was not possible to tell a float
from a fixed_array<float, 1>, which is important for Python bindings.

This commit adds an anonymous enum to DataTypeInfo with a 'Container'
value, as well as the AbstractTypeInfo::Container() method, to do just
that.


Former-commit-id: fc8d111d5e5c8ae84f8966ad303789a6fa1937e0