extlibs: define TIXML_USE_STL in tinyxml.h
With this change, we don't need to define TIXML_USE_STL each time we
include tinyxml.h.

This is a very reasonnable thing to do since it is recommended by the
documentation of tinyxml*, and it is exactly what is done with the
version of tinyxml shipped on Debian and ArchLinux anyway.


Former-commit-id: 9a250f3365fbd2d1c1fe15362390253e857bbcef