extlibs: define TIXML_USE_STL in tinyxml.h
With this change, we don't need to define TIXML_USE_STL each time we
include tinyxml.h.

This is a very reasonnable thing to do since it is recommended by the
documentation of tinyxml*, and it is exactly what is done with the
version of tinyxml shipped on Debian and ArchLinux anyway.


Former-commit-id: 9907ba71c3a8e96d1ee4fe6e18bf7dd5f4ea07bd