gui: change the directory where configuration files are saved
This commit adds a BaseGUI::getConfigDirectoryPath() method that returns
the path to the configuration directory.  (Not sure this is the best
place to put this method, but I can't think of a better place.)  It is
meant to be used when an application wants to write a configuration
file.  The configuration directory is initialized to <build>/config.


Former-commit-id: 6b6af4e49254bf09a3ee087daba4e8d04f427e94