BoundingBox: computing the bounding box of the scene is a mess. Two similar APIs are coexisting
    /// update the given bounding box, to include this
    virtual bool addBBox(double* minBBox, double* maxBBox);
    /// Bounding Box computation method.
    virtual void computeBBox(const core::ExecParams* params);

When the bounding box is used for visualization purpose (like setting the camera), only visible objects must participate on it (that is not always the case in the actual code that explains why the camera has a strange behoavior on certain scenes). The 2 APis does not mention that (are these functions only used for visualization purpose?). Do we want to compute the bounding box for another purpose?

Commenting unused functions.


Former-commit-id: f06a776790591a3e71eecae0398dd5175459d584