cmake: add a suffix to gtest libraries
The CMakeLists.txt for gtest does not use the pre/postProject.cmake
files, thus the build is not perfectly consistent with the rest of the
project.  This commit simply adds the same suffix (depending on the
CMAKE_BUILD_TYPE) to the name of the libraries as all the other
libraries built in Sofa.


Former-commit-id: b1f68e78b6c6908b932538de2a821ce4ed5eca9c