doxygen: disable FILE_PATTERNS
As it turns out, filtering out the .cpp and the .inl from the input of
doxygen yields little improvement w.r.t. speed, but it prevents the file
browser from listing all the files, and hides documentation written in
.cpp files (e.g. in tests).

The default patterns tested by doxygen seems quite good to me.


Former-commit-id: 861562441c425cb3c2b4ca95b12a2bdab81b5c5b