doxygen: move bits of documentation to .dox files
Doxygen will also look for documentation in .dox files.  For pieces of
documentation that can live on their own (like the main page, or pages
in the "Related pages" section), this is probably less confusing than
dedicated headers files.

This would also be a nice thing to do for plugins: writing the main page
in a separate file, instead of initMyPlugin.h, which is included by
every other file...


Former-commit-id: c85887023821bc0cf1e9404297f1696339778f14