simulation: remove call to unload() in ~Simulation()
Application code is responsible for calling Simulation::unload() on the
root it created, but not for resetting the root in Simulation.  Thus if
we call also unload() in ~Simulation(), it will be called several times
on the same root, with unforeseen consequences.  (A comment in
Simulation::unload() says "After this call the pointer is invalid".)

Former-commit-id: dbc410c5bfaa73eaec08954e27b16ed3f7f536a3