tests: use our own main() function
Instead of being linked against gtest_main, which contains a minimal
main() function to run the tests, the tests are now linked against
SofaGTestMain, which contains a main() function that also takes care of
initialising Sofa.

I put SofaGTestMain in tools/ for lack of a better place.

Also, this commit removes the SOFA-MISC_BUILD_GTEST build option,
because, now that I think about it, it doesn't make sense to use an
"external" gtest library.


Former-commit-id: f18e737619b84ca47d71c258f9101944efa7413f