framework/simulation: introduce cleanup() functions
The cleanup() function is the little sister of the init() function.  The
cleanup function of a library should be called before exiting to give
this library a chance to clean up, e.g. by freeing whatever resources it
may have allocated.

Basically, this means that a typical SOFA-based application looks like
this :

    int main()
    {
        sofa::simulation::tree::init();

        ...

        sofa::simulation::tree::cleanup();
        return 0;
    }

Also, this commit moves those init() and cleanup() functions to their
own files, to avoid polluting files that are included by *everything*.


Former-commit-id: a881068a9e01a81f6128228ed82140bf725a6a06