Revert "Fixing crash for multilinks."
This reverts commit e5f97bd6555c2915e6249ae13461aae7eb3cc991 [formerly 7bbc8ba4fe6a8a72369d49873b6e228fd1d53f58].

The reverted commit clears MultiLinks when calling Link::read(), which
somehow leads to NULL pointers in Nodes, with unpleasant consequences.
(E.g: launch runSofa, edit a property of a Node, click "Update", and
witness unpleasantness.)

This is what happens: the content of a Node is stored in Node.object,
which is a MultiLink.  So if we clear it, the resolution of links done
by Node::findLinkDestClass() is doomed to fail and return NULL.


Former-commit-id: dd8eb524cbc55864450d846a01e9bae8bfa203ab