core: Fix Link::read() duplicating the content of MultiLinks
The read() method reads a list of object paths, finds the corresponding
objects, and adds them to the Link.  In the case of MultiLinks, it would
add the same objects at each call.

We cannot simply remove all the objects before updating the Link,
because finding the objects might depend (indirectly) on the content of
the Link we are updating.

Instead, this commit changes the read() method to find all the objects
and store them in a list first, and only then add the missing ones, and
removes the superflous ones.


Former-commit-id: 264bcfa08f3c532ac986647f3ec68bcf2b8b698d