FIX Quater.inl add the method quatToRotationVector that implements the old toEulerVector method (changed in commit bad22c).
Calls to toEulerVector were replaced by the calls to quatToRotationVector
in files which used toEulerVector to obtain the rotation vector.
In Quater.h the  method angularDisplacement now calls quatToRotationVector
instead of toEulerVector as it should be.
In Quater.h  in the toEulerVector method add a line to avoid singularities
in the calculation (not done in previous commit 2efc4c).


Former-commit-id: 9c88c87840f3bd11ac299355bd87b71a3d128329