WARNING: Implement thread-safe Data engine/links updates by adding a lock while calling update().
This requires that requestUpdate() or requestUpdateIfDirty() be called instead of update() directly, which is now protected.
Bonus side effect is that it no longer matter where or if engines call cleanDirty() in their update method.
As a result, calls to cleanDirty() outside of requestUpdate()/requestUpdateIfDirty() are detected and a warning message is output to help update the code.
When dirty flags do need to be set outside of update(), the new forceCleanDirty() method can be used. It will not generate warnings but which may introduce thread-safety violations.