All method of the MessageDispatcher are now static.
The previous implementation was mixing part of static method with non static parts. But in both case using hidden static objects in the .cpp file.
This was totally inconsistent..so now everything is static.

An improved version would be to make a fully non-static API. But this something left for the future.


Former-commit-id: 1f8c6a3767d0858a678088d2964e2146cca4412c