Make the -z clang work properly.
By starting runSofa like this:
runSofa -z clang

The message are send to std::cerr and formatted in the gcc/clang error format that is compatible with most of the IDE.

In qtcreator in the "projets" panel you have to add a custom build step in which you start sofa (or the tests or both).
So each time you successfully build the binary, the tests and sofa are started...and their output interpreted by qtcreator.

D.


Former-commit-id: a89acb4ab58127c106567c7f7af708654fa53cb3