reverted commit f48415ef7b4152af95d6c07d2555c4ba6b908199 [formerly 87a8fccf44f03c5dc3334cd9f5676500c2937cd1]
This commit is modifying the definition of FixedConstraint, and should not be here.

The idea is to impose positions (and cancelling velocities), there is certainly an existing component to do so.
Have a look to LinearMovementConstraint and SkeletalMotionConstraint.

Worst case scenario, if nothing does exactly what you need, you can create a new component derivating from FixedConstraint.
In that case, note that the 'projectPosition' function should no touch to velocity, 'projectVelocity' is there for that.

More important, each new component in sofa-public MUST come with comments, a simple example (in Sofa/examples/Components/) and a unit test.

Thank you.


Former-commit-id: 9d6e4600921331e0468ac5c6caeea012bbfe207c