Revert : "fixed FixedConstraint & PartialFixedConstraint for size-varying mechanical objects".
Not the best way to handle this, probably bugged, and in all cases it broke projectJacobianMatrix in FixedConstraint, at least
with the implementation of constraint Jacobians using sofa::defaulttype::MapMapSparseMatrix.
In this implementation row.size() is not equal to the size of the mechanicalObject, but to the number of dofs that are affected by a given constraint equation.

This partially reverts commit c4aa645fd431369db795e317cd05df6f86460515