LiveCoding support for Python.
Monitor the files to recompiled them at runtime.
To activate this function you need to add the autoreload attribute to the PythonScriptController as in:
<PythonScriptController filename="ExampleController.py" classname="ExampleController" autoreload="1"/>

When the file is changed it is reloaded and associated to the old objects. Give it a try in the draw function.
If the python controller implements the onRecompile method this method is also called once.


Former-commit-id: 68b0b899fbaa8c6229b3d52ee2cb82770e0d5340