[SofaKernel] ADD a new type of message for 'advice'
The idea is that in some time people are using attributes or scene organization that are not considered
as good practice and should be avoided.

With this message, the component's writers can inform the users and provide suggestion on
how they should design their scene to improve their Sofa User Experience :)

Displaying of the advices should be disabled by default or controlled by the User interface (GUI or console)
but this is not implemented.

This ADD shouldn't break anything.