[SofaKernel] Add a new message formatter that may be use as default [Change the default one].
This commit changes the default formatter for testing purpose. In case of bad feedback we will
restore the old one and let this new one as optional.

Implement an improved message formatter which product much readable error message when
used in the console.

In your error message send with msg_* API you can now use the following markdown comment:
	- A double space insert a line break (you don't need to put \n or endl in your source.
	- It is possible to emphasize a text with like *this*.
	- It is possible to emphasize a text with 'this'.
	- The long paragraphs are automatically wrapped to the size of the windows console.
	- The multi-line paragraph are now nicely aligned.

In the future I would like to add URL support with [] and list item with '-'.

The general idea is that the messages emitted from the source code should be really informative
(ideally pointing to the wiki documentation pages or sofa examples).

And that this informations should be encoded in a way that is independant to where the
information is displayed (a console, a Qt Popup text area,  an html in qtSofaQuick). With this
in mind using a 'markdown' langage for the message seems a good compromise between readability in
the source code and rich features.