[SofaKernel&UI] The RichConsoleStyleMessageFormatter is now the default.
This will change how the message are formatted for anyone using the
default formatter in their application.

A second commit will force runSofa to work with the old sofa style by defautl
(not using this one)

In case this cause trouble in your application you can still use the DefaultStyleFormatter
by specifying the right formatter at the initialization of your application using the
MessageDispatcher/MessageHandler api.