[SofaKernel] FIX the broken attribute doUpdate
The doUpdate attributes appears to be broken because the selection is updated whatever
the value of doUpdate is.

This was discussed in https://github.com/sofa-framework/sofa/issues/38

This commit fix this. To stay consistent with the buggy behavior of the previous
version the doUpdate attribute is set to true by default.