[SofaKernel] BREAKING update the logging/message.h to use shared_ptr.
Replace the old 'pointer' based with the use of shared_ptr for FileInfo.

If your source code is using the messaging API (not relying on the macros)
you may needs to fix it to take that the change into account.

eg of change to do:
     aMessage.fileInfo().line will become aMessage.fileInfo()->line
     aMessage.fileInfo().filename will become aMessage.fileInfo()->filename

To avoid having to test if fileInfo() is empty or not a static FileInfo::SPtr() named
EmptyFileInfo has been created and is used wherever it is possible.