[SofaKernel] ADD FileMonitoring API + test.
A file monitoring API hiding the details of sytem specific approach.
Currently only the linux version is implemented.
A test is provided if you want to make the macosX and windows version.

This API will be used to implement:
  - python and GLSL live-coding.
  - automatic reload of data files on changes.

The API is event based and can be disabled leading to a
zero cost in term of simulation time.

TODO:
   - study thread safetiness.
   - make macos & windows version (possibly using a multi-plateform library)