[SofaKernel] ADD a class HeartBeatEvent.
It is convenient to be able trigger component ations at regular interval
even if we are in simulation mode.

This can be done by this sending this event to the scenegraph.
The heartbeat time interval is let to the application as it depend
on their specific needs and timming API. The HearBeatEvent may as
well not be emitted at all.

TODO: think about alternative way to do that.