[runSofa] ADD a new command line argument.
Live coding and some interactive python script rely on some event
traversing the complete scene regularly (heart beat event is send each 50ms,
while mouse move is sent between 125 event per second to 1000 thousand (this
depend on the mouse)).

This amount of event may be a problem for scenes that have a very large number of
object. So by default these event are not sent.

To enable them you need to use the 'interactive' command line argument.

When used the argument display a message saying this is temporary (until a better approach to
message dispatching is implemented).

TODO:
   - think about other way to implement these kind of event control as:
     - having a "masking" component that we could put anywhere and the scene and blocking the undesirable events.
     - habing an event manager...the only one receiving some event and dispatching them only to the registered component.