[Script] ADD Data comment generator
This script is a WIP to search for each initData() call and copy the "help" argument after the Data declaration in corresponding .h file with an inline '///<' comment (Doxygen readable).
It is also able to fix existing inline Data comments '//' to '///<'.

Others ways of doing this: use the ObjectFactory of SOFA, use Clang.