[SofaKernel] QUIRK-FIX linking problem that happens in release with gcc and the messagedispatcher.
When linked with gcc in release there is a missing function at link.

The problem seems to be related to the long standing dependance between Base.h and
MessageDispatcher. Some functions calls (eg: Base::getClassName) are not optimized
(and thus link succeed while other (eg: Base::getName) are somehow removed during
optimization...and thus link succeed.

I remove the use of getName to make it link again...but fundamentally this function
should be refactor to get rid of the link to Base object in the helper library.