[applications] DONE-Renaming the heartbeatevent into idleevent.
This version should compile and work nicely.
Warning this version may break any component that were using the HeartBeatEvent
(but I suspect no-one is currently doing that yet).

This version breaks the scripts that are using the onHeartBeat function. Those have to be updated to overload the
onIdle function instead.