[SofaKernel & SofaModule] CLEAN add TODO about unifying coloring scheme.
This commit show where colors are passed to sofa component... it appears
that in 2/3 of the case only 1,0,0 syntax is allowed while the others also
allows "red" "blue" color definition.

This is an UI consistency issue. Because of UI consistencies our users
think that sofa is:
- hard to understand & master
- not well done

So we should unify that.
So I add a TODO where this should be done and maybe I will fix some of them :)