[SofaOpenGLVisual] Add new to to ClipPlane visual component.
There is two aspect I think should be considered as open issue.
One is that a data field called "normal" is not warning the user that a warning
should be normalized (or it should normalized it by itself)

The second is that the id is not checked for negative values which is probably
a bug.