[SofaKernel] ADD data fields for fine grain control of the message log.
'messageLevel' should be use to specify above which level a message is emitted by the component.
'loggingLevel' should be use to specify above which level a message is loggued in the component.
'loggingSize' should be use to specify the size of the internal logging queue.

I'm not fully convinced of that.