[SofaKernel] WIP Fixing the compilation problem on win32
The compilation problem with visual 2015 seems related to template
instantiation and the way the compiler handles static_assert in
fixed_array.

I have no idea how to fix that so I passed the object into a
template version. If this solve the issue this will be definitely the most
un-satisfactory commit of my all life uglyfing a very readable and clear
implementation to turn it into a messy useless templates.

I secretely hope this commit will not fix the problem :)