[All] Remove the old output & warning buffers all around.
the old string based buffer are now replaced with a proper API
to log & access message in component.

This new API is now used in the runSofa GUI to improve
the messages in the panel.

The Output & Warning panels have been removed into a single one
because it is much better (I can explain but I'm too tired).

I remove the clearWarning and clearOutput from some visitors...
Again I could explain but I'm tired.