[SofaPython] A reinit approach based on a scene component
The drawback of the previous approach, let's named it 'reinit' approach is that it add
a new concept to the PluginManager api.

An alternative implémentation is provided here that adds a
new component to each scene. This component is in charge of
saving/restoring the python context when the scene is loaded/unloaded.
The semantic fit quite nicely but this have a complexity cost because there
is no way to install a component once per scene so we have to
search if it exists each time with the getObject() function.