[SofaTest] CLEAN to prepare for a PR
Removes the _2 suffix that remains.
Seprate the .h in two part, the public and the private one.
The private one is moved into .cpp.
The public one is documented in kept in the .h.

Also remove some 'no really' needed piece of code.
The implementation is far from perfect...but it should be ok.