CHANGE usage of d_totalMass in DiagonalMass and add tests
This commit sets the d_totalMass as read and write in DiagonalMass.
Before set as read-only, without reason.
The commit also updates the coding rules in the Data.

Tests are added to check for both hexa and tetra topologies,
the coherency between totalMass and massDensity. It also checks that
if both totalMass and massDensity are defined, totalMass predomines.

WARNING: the test showed issue of casting in the EXPECT_EQ()