[SofaKernel] Fix getRelativePath() in DataFileName + add tests
Problem:
The implementation of getRelativePath is broken as it returns the current
value which can be either a relative or an absolute path.

Solution:
The new implementation is building a real relative path against the path contained
in the DataRepository.

I also added few test for DataFileName to valide the different behaviors.