[SofaKernel] CHANGE behavior of FileRepository::getRelativePath
The WIN32 implementation is lowering the case of the given path. This is
not consistent with other case insensitive OS like macos.

In this PR we deprecate this 'lowering' behavior, adds a dedicated message & update
the call point all around Sofa. Existing code should continue working but a
message will be printed to indicate to user they must update their code.