[MeshROI_test] Fix test failure/crash un MeshROI_test
The tests where not really working as they were 'forcing' the use of given
positions while using a topology for the edges & triangles implying more indices.

Because of the too large indices the MeshROI object was thus crashing at init().
I changed the test so that they are using a valid behavior based on a consistant topology.

todo: add a new test to demonstrate the problem.