[SofaCUDA] FIX: NVCC flags for debug build on Windows
Targets are built with debug version of runtime library in debug configuration
under MS Visual Studio. This is enabled with /MDd compiler flag. Missing this
flag for NVCC generated object files results in linker fails due to different
versions of runtime libraries.

The problem is resolved by explicitly setting compiler options for debug build
under Windows. The solution is given only for Windows and only for MSVS because
it's the only officially supported by NVidia compiler on this platform.