[SceneCreator] Refactoring SceneCreator (minimal breaking)
The scene creator implementation was using the real c++ types to create
the object which was problematic in term of depency to manage to pluginize
sofa. To cut the dependencies we have implemented a new api in
sofa::simplapi:: this API helps to create scenes using the factory and strings
in a way similar to what python is doing. To make it more abstract
I also changed the SceneCreator.h signature to take BaseObject instead
of MechanicalObject.

The SceneCreator.h was also having some function we can wonder how they are related
to the creation of scene. So I move them into SceneUtils.h

Finally I propagate the changes to the files using SceneCreator.