[SofaKernel] config.h refactoring.
Currently config.h is a all-in-one configuration file that is generated by
CMakeLists and included nearly everywhere. The problem is that it contains
also very specific options that very files are using, but, each time the option
is changed all sofa needs to be recompiled. This is why I suggest to switch to a
more specific approach in which .cpp that specifically requires a given define
load the corresponding config.
Eg:
#include<config/build_option_opengl.h>  // to do #if(SOFA_WITH_OPENGL==1)
or
#include<config/build_option_experimental.h>  // to do #if(SOFA_WITH_EXPERIMENTALFATURE==1)
or
#include<config/sharedlibrary_defines.h>

... it is still possible to use the old file but by being more specific we will
reduce the amount of file to recompile when we will changed something.