[SofaKernel] config.h refactoring.
Currently config.h is a all-in-one configuration file that is generated by
CMakeLists and included nearly everywhere. The problem is that it contains
also very specific options that very few files are using. But, each time any of these specific
option is changed sofa needs to be recompiled entirely.

This PR is a implementing a more specific approach in which each .cpp that
specifically requires a given define a dedicated file.
Eg:
#include<config/build_option_opengl.h>  // to do #if(SOFA_WITH_OPENGL==1)
or
#include<config/build_option_experimental.h>  // to do #if(SOFA_WITH_EXPERIMENTALFATURE==1)
or
#include<config/sharedlibrary_defines.h>

... it is still possible to use the old file but by being more specific we will
reduce the amount of file to recompile when we will changed something.