[SofaKernel/frameworkextra_test] Fix test to correctly load the required plugin
NB: There is a very interesting issue here about the required-plugin and
compilation links. It seems non trivial here to guarantee a full consistency
between the tests executed and the plugin states (i.e garanteeing that the
plugin is correctly reconstructed if the source code changed).

It is yet unclear to me how to handle that but this may be taken into account for the SofaNG.