[CMakeLists.txt] Add macro to handle out-of-tree plugins.
Currently there is the SOFA_EXTERNAL_DIRECTORY to add out-of-tree
directory in the compilation process. But when this directory contains
plugins those one are not handled the same way as are the one in
application/plugin.

To unify that I added a SOFA_EXTERNAL_PLUGIN option and
did minimal change to the 'sofa_add_generic' macro to detect and
handle out of tree directory.

So if you want to add external plugin you can use SOFA_EXTERNAL_PLUGIN
instead of SOFA_EXTERNAL_DIRECTORY.