[StentExp] ADDS new plugin base files
Creation of a new plugin StentExp, which aims at simulating a ballon
expandable stent during its expansion.

The idea is to model the linear parts of the stent using beam elements,
close to the BeamFEMForceField component, already implemented. The newly
created MultiBeamForceField component is currently identical to the
BeamFEMForceField component, except for the beam element section.

The beam that was previously considered as a cylinder (or a hollow
cylinder) now has a rectangular section. The second moment of inertia
and the visual display are modified accordingly.