[runSofa] Add a DocBrowser window to view in better condition the docs.
Refactor the way the html documentation associated with Sofa scene
are handled. The doc browser now have an history and link to sofa
scene are correctly handled to load the sofa scene. This is very
practical to make self documenting libraries.

The DocBrowser can be displayed from the Help menu.