[Flexible, image, CImg, CImgPlugin] Refactor to FIX crashes.
Problem:
CImg is a .h only library that contains a mutex to allow multi-threading
use of the library. The problem is that when the CImg/CImg.h is included
in several different shared library (i.e. plugin) then it is unclear
where the mutex are and when they are initialized. This scenario is visible
in Flexible where there is scenes (demos/steak.scn) that use both image
& CImgPlugin (as CImgPlugin is now a plugin & mandatory).

To fix that I refactor how CImg is included in the software.
The idea is to a have a single definition of CImg's mutex in
initCImgPlugin.cpp (so in CImgPlugin.so) and have all other
libraries that uses CImg to:
find_package(CImgPlugin REQUIRED) (instead of CImg)
then
target_link_libraries(${PROJECT_NAME} XX YYY ZZ CImgPlugin)

and in .h do:

[CImgPlugin] Clean a bit the whole CMakeLists.txt.

- CImgPlugin.h.in is useless (there is no value expanded)
- the test is now externatlized from the plugin (with correct dependencies to it)
- Remove useless header file inclusion.
- Remove commented code

[CimgPlugin] Centos WIP ?

[CImgPlugin] (WIP) Try to force the linking of the mutex for CImg.

[image] Trying to fix the link issues with CImgPlugin.

[Cimg] how to export properly the symbol on windows ?

[examples] Fix example that were using drawColorTetrahedra with three values instead of four.

[CImg] MOVE extlibs/CImg into extlibs/CImgPlugin

It is not allowed to used CImg directly anymore as having multiple
instances cause crashes...use SOFACImg instead.

[CImgPlugin/CImg] Update CMakeLists.txt to handle the move of CImg.