[CImgPlugin, Flexible, image, CImg] Refactor CImg and move it into CImgPlugin then fix the code that depends on it.
The general problem is that CImg is header only library thus the
multiple inclusion of the code in multiple dll seems to cause trouble
on the mutexes initialization causing the crash & failure when two plugins compile
with CImg & CImgPlugin are loaded (eg image, flexible).

To avoid this scenario, CImg is moved into the CImgPlugin that
hold the mutex and other code needs to include #include<CImgPlugin/SOFACImg.h>
instead of the raw CImg.