[SofaGraphComponent] FIX RequiredPlugin
The plugins was allowing either to provide the plugin to load
in the component's name or in the pluginName attributes. To increase the
consistency if the plugin to load is provided in the name attribute then
the 'name' is copied in the pluginName lists.

So that code using RequiredPlugin knows that the list of plugins are in
pluginName.