[plugins/MultiThreading] FIX problem at initialization of mTaskMutex.v_
The plugin is broken since PR #549 as it removes the initialization
of the mutex. Possible explaination is that v_ is implementation
dependent but the existing initialization was done incorrectly
using a (int).

Now use BOOST_DETAIL_SPINLOCK_INIT which insure that the initialization
is coherent with the implementation used.

(I also cannot prevent myself to remove the two ugly commented lines)