CHANGE initialization process and datanames in UniformMass
Unify the data names :
- vertexMass corresponds to a single real value defining
the mass set at each vertex (previously called mass)
- totalMass corresponds to the total mass of the object
(previously called totalmass)

Implement a clean and rigorous initialization process:
first the vertexMass is checked. If set by the user, it
will be used for initialization and totalMass is computed
from it.

If vertexMass is undefined or wrongly set, the totalMass
will be used. If it is set by the user, the totalMass will
be used for initialization and vertexMass is computed from it.
If totalMass is undefined or wrongly set, the default value
of totalMass = 1.0 is used, and vertexMass is computed
accordingly.

New tests added and file for rigid case updated with the
new name vertexMass instead of mass.