Unify the initialization process in DiagonalMass
Implement a clean and rigorous initialization process:
first the vertexMass is checked. If set by the user, it
will be used for initialization, massDensity and totalMass
are computed from it.

Second, the massDensity is checked. If set by the user, it
will be used for initialization, vertexMass and totalMass are
computed from it.

If other infos are undefined or wrongly set, the totalMass
will be used. If it is set by the user, the totalMass will
be used for initialization and vertexMass is computed from it.
If totalMass is undefined or wrongly set, the default value
of totalMass = 1.0 is used, and vertexMass is computed
accordingly.

New tests added for different initialization cases