[SofaHelper] FIX readFromIniFile plugin loading
m_pluginMap[pluginPath].initExternalModule(); // this line was wrong

This line was creating entries in m_pluginMap because pluginPath is not always a path.
It is not a path when we are reading "plugin_list.conf.default" file to auto-load plugins at startup.

+ ADD getPlugin method to avoid future operator[] mishandlings.
Get a loaded plugin by paths OR names (even if the map is based on paths).