[StentExp] ADD virtual displacement method
The previously implemented method used in the MultiBeamForceField
component to compute the stiffness matrix was relying on the virtual
force method as described by Andersen and Nielsen in their "Elastic
Beams in 3D" lecture notes.
Here we implement another general method for finite element computation
- described for instance in Zienkiewicz and Taylor's "The finite Element
Method", Butterworth Heinemann - which

The second major change is the introduction of plasticity. Two types of
plastic models are implemented: the first one is related to the first
(virtual force) method, and the second one to the virtual displacement
method.
The virtual force method plasticity is potentially irrelevant. The
test of the plastic threshold is made with a strain quantity which we
compute artificially from the nodal forces distributions. It is then
likely that this computation has no physical meaning. Tests made on
single beam elements show divergent behaviour.
On the other hand, the plastic behaviour computed from the second method
follows the ideas developped in Muller et al (2004), which have already
been implemented on tetrahedron elements in SOFA. This model corresponds
to ideal plasticity, and therefore gives somewhat unrealistic but
coherent results.