[StentExp] ADD plasticity handler class
The plasticity behaviour in the MultiBeam component is gathered inside
a specific (internal) class MultiBeamPlasticityHandler.

For the moment, this class implements the resolution method proposed in
Krabbenhoft's Basic Computational Plasticity lecture notes (2002). All
necessary intermediate functions are implemented, but the resolution
itself isn't, as the computation of the time step increments is done in
a higher level in Sofa (i.e. in the solver component).
NB: These intermediate functions have not been tested yet.

Note that the previously implemented (simple) model of plasticity is
still valid. Once the more complex model is validated, it should replace
the former.