[StentExp] CHANGE the reduced integration handling
Foreseeing the addition of a more complex plasticity model - that requires
the update of the (tangent) stiffness matrix at each time step - we pass
part of the reduced integration process to the initialisation of each
beam element.
The idea is that the Be matrices (containing the derivatives of the shape
functions) can be computed once and for all at the begining of the
simulation. As the reduced integration intervals depend on the length of
each beam, it is necessary to store the matrices in the BeamInfo
structure. We count a total of 27 Be matrices per element (one for each
of the 27 Gauss points).

NB: for the moment, the only link between the matrices and the
corresponding Gauss points is that both follow the same order in the
gaussian quadrature object and the vector of matrices. This comes from
the fact that the matrices are originally computed by a gaussian
integration procedure.
A better solution could be to ordinate the Gauss points inside a field
of the quadrature object (in an array for instance). Hence the order of
the Gauss points would be explicitly given and wouldn't depend on the
methods concretely processing the integration.