[StentExp] ADD more complex plasticity behaviour
Implementation of a beam element plasticity behaviour according to the
plasticity model described in Krabbenhoft's lecture notes(2002)

The idea is to distinguish the cases where the deformation is purely
elastic, and the cases where the deformation is plastic. This is handled
in the addForce method, which calls a different sub-function
(accumulateNonLinearForces) if the deformation is plastic.

The deformation process starts with an elastic phase, where the forces
and displacement are the same as for a purely elastic element. The
difference is that the new stress state is stored in a variable at each
time step. In addition to that, this stress is tested to check if it is
higher than some plasticity threshold (defined by the Von Mises Yield
function).
If in a time step the stress is higher than this threshold, we recompute
this stress through a plastic deformation process (as in krabbenhoft's).
This process includes an incremental stress computation (which can
involve up to 25 iterations per time step and per Gauss point), and
relies on the solution of an implicit linear system.
Once this stress increment is computed, the corresponding nodal forces
and tangent (linearised) stiffness matrix are updated, and returned by
the function as in the linear case.
At each of these plastic iterations, the newly computed stresses are
also tested, to check if the deformation has actually returned to an
elastic phase.

NB1: For the moment, the Von Mises yield function doesn't include a
hardening parameter, which would concretely make the plastic threshold
change (increase) as plastic deformations occur. This is to be
implemented for a more realistic plastic behaviour.

NB2 : The handling of the post-plastic phase should not be exactly the
same as the first elastic phase, as it should take into consideration
the plastic history. This, as well, is not yet implemented.