[StentExp] ADD plastic deformation history
The plastic history of the deformation is now stored in a field of the
BeamInfo class (_plasticStrainHistor). It contains a plastic strain
tensor computed at each Gauss point of each beam element. This tensor
corresponds to the plastic strain state when the beam returns to an
elastic phase, after plastic deformation.
We highlight the fact that the computation of the plastic strain
involves the solution of a linear system, to get the elastic strain which
would correspond to the last plastic stress state. However this
situation should only occur when passing from a plastic to an elastic
phase.

For the moment, releasing external forces does not make the beam go back
to an elastic phase. This error will be dealt with in the next commit.

NB: we also introduced an (arbitrary) threshold that prevent the element
to leave the plastic phase because of numerical approximation.